Implement Redc_Is_Fake meshes (for collision)
The idea is that by exporting meshes with a fixed accessor name for
vertices and indices, we can query the data in the engine easily,
without telling the render code to ignore some objects in particular.
This code basically checks for some mesh properties that a) signal
whether the object should be ignored for rendering (Redc_Is_Fake) and b)
designate the names for the vertex and index accessors.

There is a bit of duplication in this commit like between the functions
export_mesh and export_fake_mesh.